misc: modularize helper functions
