+++ lib/defaults.js
@@ -1,0 +1,10 @@
+module.exports = {
+  pageToken: '##gulp-resource-hints##',
+  paths: {
+    prefetch: '**/*.+(png|svg|jpg|jpeg|gif|woff|woff2)',
+    preload: '',
+    preconnect: '',
+    'dns-prefetch': '',
+    prerender: ''
+  }
+}
